Resolve race condition issues, and mark a few tests as exempt from race detection
