+++ core/internal/storage/inmemory.go
@@ -43,7 +43,8 @@ type InMemoryStorage struct {
 	minDistance int64
 
 	requestChannel chan *protocol.StorageRequest
-	running        sync.WaitGroup
+	workersRunning sync.WaitGroup
+	mainRunning    sync.WaitGroup
 	offsets        map[string]clusterOffsets
 	groupWhitelist *regexp.Regexp
 	workers        []chan *protocol.StorageRequest
@@ -86,7 +87,8 @@ func (module *InMemoryStorage) Configure(name string, configRoot string) {
 
 	module.name = name
 	module.requestChannel = make(chan *protocol.StorageRequest)
-	module.running = sync.WaitGroup{}
+	module.workersRunning = sync.WaitGroup{}
+	module.mainRunning = sync.WaitGroup{}
 	module.offsets = make(map[string]clusterOffsets)
 
 	// Set defaults for configs if needed
@@ -134,10 +136,11 @@ func (module *InMemoryStorage) Start() error {
 	module.workers = make([]chan *protocol.StorageRequest, module.numWorkers)
 	for i := 0; i < module.numWorkers; i++ {
 		module.workers[i] = make(chan *protocol.StorageRequest)
+		module.workersRunning.Add(1)
 		go module.requestWorker(i, module.workers[i])
 	}
 
-	module.running.Add(1)
+	module.mainRunning.Add(1)
 	go module.mainLoop()
 	return nil
 }
@@ -148,17 +151,18 @@ func (module *InMemoryStorage) Stop() error {
 	module.Log.Info("stopping")
 
 	close(module.requestChannel)
+	module.mainRunning.Wait()
+
 	for i := 0; i < module.numWorkers; i++ {
 		close(module.workers[i])
 	}
+	module.workersRunning.Wait()
 
-	module.running.Wait()
 	return nil
 }
 
 func (module *InMemoryStorage) requestWorker(workerNum int, requestChannel chan *protocol.StorageRequest) {
-	module.running.Add(1)
-	defer module.running.Done()
+	defer module.workersRunning.Done()
 
 	// Using a map for the request types avoids a bit of complexity below
 	var requestTypeMap = map[protocol.StorageRequestConstant]func(*protocol.StorageRequest, *zap.Logger){
@@ -193,7 +197,7 @@ func (module *InMemoryStorage) requestWorker(workerNum int, requestChannel chan 
 }
 
 func (module *InMemoryStorage) mainLoop() {
-	defer module.running.Done()
+	defer module.mainRunning.Done()
 
 	for r := range module.requestChannel {
 		switch r.RequestType {
