+++ core/internal/zookeeper/coordinator_test.go
@@ -1,3 +1,5 @@
+// +build !race
+
 /* Copyright 2017 LinkedIn Corp. Licensed under the Apache License, Version
  * 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
